road/simulation update refactores, renamed Game->Simulation, Road moved to simulation
