+++ core/simulation/__init__.py
@@ -1,3 +1,3 @@
 from core.simulation import *
 from core.simulation.car import *
-from core.simulation.game import *
+from core.simulation.simulation import *
